Work in progress on annotate command
I have given up hope that I can repair the original implementation of the Annotate function. The original version places a WinForms user control inside a document window and embeds a text editor in that user control.

So far as I can tell, it is no longer possible to place a WinForms user control directly in a document window.

The approach taken to embedding a text editor in the user control is designed to work in Visual Studio  going back to version 2005. I suspect that it is much more complicated than it needs to be in Visual Studio 2019.

For these two reasons I have started restructuring the implementation using a WPF control and instantiating the editor in a completely different manner.

This is long way from being finished, but it is the bones of a working implementation.

However, I am still considering whether the annotation margin should be implemented as an editor margin. Fundamentally I think that would be a good idea, but I still have some open questions about how to implement that.

If I went that way, it would require more restructuring.