Display context menu on annotation margin. Implement the commands "Compare to working copy" and "Show changes".
The context menu DOES NOT USE existing menu system in AnkhSVN OR the selection manager.
Using these mechanisms involves a lot of boiler plate code and in my opinion is way to complicated.

Instead, the commands are implemented as WPF-Style command objects in the AnnotateRegion class.

To access the difference service with the IAnkhDiffHandler interface, I now pass the context object
with the interface IAnkhServiceProvider through several levels into the AnnotateSource class, where
it can be accessed from the AnnotateRegion class.

(I could have stored it in AnnotateRegion instead, but I figured it was slightly more flexible to store
it in AnnotateSource and would require less storage.