Add a method to interface IAnkhTempDirManager to remove temporary directories. Override the Dispose method in AnkhSvnPackage and explicitly call this new method.
Formerly, the temporary directories were cleaned up in the finalizer of TempDirCollection,
but - at least on my system - this method is no longer called.

I have also tried to install an event handler for OnBeginShutdown in TempDirManager, so
that it can continue to clean up on its own, but I failed to get this working.