Conversion to MEF - Convert AnkhRepositoryProviderService
Convert the service AnkhRepositoryProviderService from an AnkhService to an MEF component.

This is a very simple case, which does not really merit being a service at all.

This is still not a proof of concept.

I have not managed to use the MEF container in Visual Studio and I am using the IAnkhPackage service to get the MefContainer (which is against the spirit of clearing away the AnkhService architecture).

Furthermore, I am specifying a constructor parameter "AnkhContextParameter" to pass the context into the created service, which is also rather against the spirit of these changes.