Conversion to MEF - Add MefContainer to BaseCommandEventArgs
Add the MefContainer (CompositionContainer) to BaseCommandEventArgs and initialise it, so that it is passed into the command handlers.

I don't know whether this will be a long term solution, but at present it seems practical.

Pass the MefContainer to the constructor of RepositoryProviderPage and (one level higher) RepositorySelectionWizard and call SatisfyImportsOnce in constructor of RepositoryProviderPage (rather than later).