Issue #12
There is already a mechanism in SharpSVN to specify which SVN errors should be handled by generating an exception, and which should not.

In the method PendingChangeHandler.Commit_CommitToRepository a specific list of errors is added to the class SvnCommitArgs using AddExpectedError.

I have added the error SvnErrorCode.SVN_ERR_WC_PATH_NOT_FOUND to this list.)